chore: move xcm configs to the xcm config module
